Helper Function to Convert a Vector of Parameters to a List of Parameters

Helper Function to Convert a Vector of Parameters to a List of Parameters

theta_helper(x, mean_structure = FALSE)

Arguments

x
Numeric vector. Vector of parameters \(\boldsymbol{theta} = \{ \boldsymbol{\beta}, \sigma^2, \mathrm{\vech} \left( \boldsymbol{\Sigma}_{\mathbf{X}, \mathbf{X}} \right) \}^{\prime} \) when mean_structure = FALSE or \(\boldsymbol{theta} = \{ \boldsymbol{\beta}, \sigma^2, \mathrm{\vech} \left( \boldsymbol{\Sigma}_{\mathbf{X}, \mathbf{X}} \right), \alpha, \boldsymbol{\mu}_{\mathbf{X}} \}^{\prime} \) when mean_structure = TRUE

mean_structure
Logical. If mean_structure = TRUE, x includes mean structure parameters. If mean_structure = FALSE, x only includes covariance structure parameters.

Value

A numeric vector.

Dependencies

rmvn_chol() (test)

See also

Author

Ivan Jacob Agaloos Pesigan